feat(desktop): relay management-API calls through the Rust host
The volunteer daemon's local management API now rejects requests whose
Host header is not its own loopback address and sets no CORS headers, so
the app's web view can no longer call it with a browser fetch: every call
failed at the preflight. Calls made from the Rust host process succeed.

Add a generic `mgmt_request` Tauri command (method, path, optional JSON
body) that reads ~/.lettuce/daemon.json on every call, sends the request
with the bearer token, returns the parsed JSON (null for 204) and maps a
non-2xx response to a structured {code, message, status} error taken from
the daemon's error envelope (UNKNOWN when it cannot be parsed,
DAEMON_UNREACHABLE when no request reached the daemon). Requests time out
after 15 s, or 30 s for /api/v1/credit, which asks every attached head in
turn. The typed helpers used by the tray, notifications and container
runtime code share the same transport. `get_daemon_info` is removed: the
web view no longer needs the port or token.

Harden the structs other Rust modules deserialize (status, metrics,
container runtime, notification preferences) so absent fields default
instead of failing the parse, and read credit as a decimal now that heads
report fractional credit.

Add three host-side commands the UI needs: `system_metrics` (host CPU and
memory via sysinfo, with one cached System so CPU usage is measured
between calls; the daemon reports zeros for these), `restart_daemon`
(graceful stop, forced after 30 s, then start and wait for a fresh
daemon.json with a live PID) and `get_client_version`
(`lettuce-volunteer --version`). `run_init` now passes
`--trust <container,native|none>` whenever it passes `--server`,
mirroring the CLI's accepted values, and drops the unused schedule hour
fields.

+++ apps/volunteer-desktop/src-tauri/src/container_runtime.rs
@@ -1,7 +1,11 @@
 use crate::api::{DaemonInfo, ManagementClient};
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+/// `GET /api/v1/container-runtime`. Every field defaults when absent; the
+/// machine sizes are integers on the daemon side but are read as i64 so a
+/// future change in width cannot fail deserialization.
+#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+#[serde(default)]
 pub struct ContainerRuntimeStatus {
     pub backend: String,
     pub status: String,
@@ -9,9 +13,9 @@ pub struct ContainerRuntimeStatus {
     pub socket_path: String,
     pub machine_required: bool,
     pub machine_name: String,
-    pub machine_cpus: i32,
-    pub machine_memory_mb: i32,
-    pub machine_disk_gb: i32,
+    pub machine_cpus: i64,
+    pub machine_memory_mb: i64,
+    pub machine_disk_gb: i64,
     pub error: Option<String>,
 }
 
@@ -22,7 +26,8 @@ pub struct SetupRequest {
     pub disk_gb: Option<i32>,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, Default)]
+#[serde(default)]
 pub struct SetupResponse {
     pub status: String,
     pub message: String,
